Add in related links to contacts presenter
This one has a few smelly bits, specifically having to use indifferent_access
the raeson we need to do this, is the templates expect symbols, and the json
that comes back is not in symbols. My recommendation here is to
make the tempaltes deal with this in the background as much as possible.
Also had to massage related_links, as they do not have urls, instead web_urls